Tweaks to the LSUIElement stuff.
Makes revealing from the background silent because TransformProcessType gets the thing done at runtime. Also only makes it change if the setting could be applied to the Info.plist file.

NSApplicationPresentationOptions only allows an application to change the state of interface objects on the system (Showing/Hiding the Dock, …) so it's not of any use to us (I'm removing the warning because of it).

As per http://codesorcery.net/2008/02/06/feature-requests-versus-the-right-way-to-do-it, the current approach will break when Apple forces code signing onto us all...